added test for DataProvider mutability
sometimes you need to mutate dataProvider (for example when you have to
set up a new session token in the HTTP header), in that case we make
sure that no error is thrown.

Previously there was a bug where you would get an error regarding
apiSchema not being set if you mutate dataProvider whithout reloading
the API Schema. This test is for preventing a regression.